Rename project to FC4C
I’m giving this overall approach a name: FC4. I’ll document the reasons
and meaning of that specific name in the upcoming FC4 repo. I was
considering naming this tool FC4 as well — I want a short name for the
tool so it’s easier to use — but I decided that would be too confusing
and ambiguous. So I added a C. It doesn’t mean anything, and I only
chose the C for the way it all sounds. It’s a slight nod to R2-D2, I
think, just with its cadence.