clipboard: replace get fn with clipboard var
I just happened to be in this file and I noticed that this didn’t need
to be a function and that the code would be more readable and more
efficient if this was just a value in a var rather than a function.

I’m not 100% sure if this is idiomatic, but I like it.