clipboard: check content type before reading
I noticed occasional errors when something other than a string was in my
clipboard, for example the stacktrace in #12.

This might in fact mostly address #12, but looking at the code in repl/wcb it looks like to *truly* address #12 the entire body of the go-loop needs to be wrapped in a try/catch.