Refer to diagrams as FC4 rather than Structurizr [Express]
Now that we’ve documented the FC4 framework as its own thing with a name, it makes sense for this tool to refer to it rather than directly to Structurizr Express (SE).

Note, though, that the framework does still use SE, and our data structure used for defining diagrams is still the one defined by and used by SE. This does, however, lay the groundwork for eventually diverging from SE, which I suspect we’ll probably want to do at some point — maybe sooner rather than later. Among other problems, SE strips out comments and any properties it doesn’t recognize from the YAML, which is very limiting for our purposes.

I also took the opportunity to streamline some function names. I realized that what in the code and docs I had been referring to as “a parsed Structurizr doc” is really just a diagram as a map, i.e. a map representing a diagram. Once the diagram has been parsed from the YAML, it’s no longer a “doc” and the fact that it had at one point been serialized as YAML is irrelevant at that point.