Some Tests! 🎉 (#19)
This is a squash-merge of 19 commits:

* Improve some specs

Basically I ran the tests a bunch of times and saw some failures that
were not due to errors in the functions under test but rather errors in
the specs -- some of them were generating unrealistic values. These
changes should prevent those failures.

* Reduce the number of checks of a generative test

One of these function specs that’s being checked, shrink, it generates
massive values, and also the predicates that make up the spec are not
fast. So this reduces the number of checks run of shrink to a
still-reasonable 300.

* WIP Create diagram spec

* Tests!

* Added another function spec

* Rename test files to match src files

src files were renamed way back in ac316538 but this work to add tests
is currently in a long-running branch... tsk tsk.

* Update test runner and trifurcate aliases

* WIP move specs to dedicated file with multiple namespaces

* WIP PROGRESS

* MORE WIP on spec-based testing

* MORE WIP on spec-based testing

* cleanup whoops!

* All function specs now passing tests!

* shrink should preserve types

* Remove shutdown-agents hack

Doesn’t seem to be needed anymore

* Clean up and simplify specs

Realized that now that we’re using req-un and opt-un, we won’t need to
qualify the keys in the diagram maps, and therefore we don’t need the
“convenience” specs that were really just pointers to more generic
specs.

I thought I was going to need those “convenience” specs because I had
forgotten about req-un and opt-un, so I had been thinking that I’d need
to process each diagram map as parsed from the YAML, with unqualified
keys, in order to qualify the keys. I had been thinking that it’d be
really hard to know which namespace to apply to each key, unless they
were extremely regular — i.e. if every map used the same namespace for
all its keys.

Anyway, turns out, that was all wrong; tldr Rich Hickey is a little
smarter than me.

* Add a function spec

* Fix a spec

We have some functions that round up or down. So this generator would
generate some very large but valid value like 49997 and it would get
rounded up to 50000 — which was invalid, as the upper limit here is
exclusive. This change should prevent that.