Use clojure rather than cli when running tests
The clj command failed in the CircleCI build because the Docker image
circleci/clojure:tools-deps doesn’t have rlwrap installed.

So I checked out the reference for tools.deps and found:

> The clojure tool is a runner for Clojure.
> clj is a wrapper for interactive repl use.

https://clojure.org/reference/deps_and_cli#_usage

I also updated the ENTRYPOINT in the Dockerfile to use `clojure` rather
than `clj` because (a) now we don’t have to install rlwrap, and (b) now
we don’t have to have a weird sleep workaround in there to work around
some issue with rlwrap and Docker.

In these aspects:

* Use the `clojure` script rather than `clj`
* Use the `-e` option to specify the form to compile, rather than
  piping it in via stdin

This is easier to read and understand and also more consistent with the
README and with .circleci/config.yaml