[WIP » fc4c.model] Start: new home-grown data model
It comprises a model, a set of views, and a set of styles.

Background from the (private within Funding Circle) ticket:

Currently with FC4, we have only a single kind of data structure: a
diagram. It defines both the systems and their elements and all the
relationships thereof, and how those should all be depicted in a
specific view.

This is very straightforward in the sense that we end up with a 1:1
correlation between YAML file and PNG file, and we have only one kind of
file/structure: a diagram.

But! It has a bunch of downsides:

* Requires lots of duplication and repetition
  * e.g. a container diagram is a superset of the same system’s context
    diagram
  * e.g. the same element may appear in multiple files, and its content
    (data) (e.g. description) need to be repeated each time
* That duplication and repetition leads to too much busywork to keep
  things in sync, and too much risk that things will drift

Really, each thing should be defined/described/stated in one single
time+place. A “single source of truth” — look at that, that idea pops up
_everywhere!_

So let’s try embracing one of the aspects of
[Structurizr|https://structurizr.com/] that I had initially dismissed:
defining the _model_ of one’s systems independently of the _views_ of
those systems.

Concretely, this would mean that instead of a single dir tree containing
simply _diagrams_, we’d have a bifurcated tree containing the _model_
and the _views_ (or perhaps we should simply call the views _diagrams_).