Improve fc4c.files/relativize error behavior
Generative testing wherein directory path strings are generated randomly
via the specs in fc4c.model revealed that this function was expecting
that the parent *will* contain the path. That makes sense because that’s
the whole point of this function, to help relativize a path in this
case. That said, it’s better for a function to be more resilient, and to
have more graceful failure modes than just throwing an exception. So now
the function just returns nil if the path is not a child of the parent
(i.e. if the parent path is not included in the child path).