[WIP++ » fc4c.model] Run cljfmt on model.clj
I used the default settings for cljfmt. I don’t know if these are
what we’ll want long term; for now I just wanted to improve the
consistency within this new long file.

I plan to run it on all the other files in this project in a subsequent
branch with its own PR.

Also:

* includes one fix it made to files.clj, since that function had
  already been changed in this branch
* adds two new profiles to deps.edn: lint and lint/fix