Fix name of dependency in deps.edn
My impression from reading the tools.deps
[guide](https://clojure.org/guides/deps_and_cli) and
[reference](https://clojure.org/reference/deps_and_cli) is that the
name of the dependency (which is used as the map key) does’t _really_
matter when using git coordinates (unlike with Maven, wherein the lib
name is the group name and artifact name, joined by `/`). As far as I
can tell, they really just need to be unique within their map, because
map keys must be unique.

That said, using the same name as the forked dependency was confusing;
it lead to one reviewer seeing this, noticing that the sha values were
different between this and the reference to com.cognitect/test-runner,
and pointing that out as a potential problem. Which was a good catch and
totally reasonable; I think using the same identifier/name/key for two
different libraries in this file was a mistake.

With this new name I’ve followed the convention suggested by the
tools-deps guide.