Replace tag :in-house with :internal in model
I realized I’d been using :in-house because the tag `internal` has a
special meaning to Structurizr Express (SE), so it can’t be used
generally in an SE diagram. This model of a system landscape,
though, is intended to be decoupled from and independent of SE. So
within this data model there’s no reason _not_ to use `internal` and
there *is* a good reason *to* use it: it’s more intuitive.

The forthcoming ns that exports FC4 views to SE diagrams can and
probably should transform :internal to :in-house automatically.