Fix test that always failed when measuring coverage
I have zero clues as to why, but before this change the generative tests
for snap-coords, which uses round-to-closest, would always fail when
running the tests with coverage measurement enabled, i.e. from within
Cloverage. This, despite that they always passed when run from the REPL
or from the standard test runner runner.

Anyway, this means that the tests will cease failing when measuring
coverage, which means we can fail builds if Cloverage returns a non-zero
exit code, and we can maybe even eliminate the redundant test runs from
the CI build job.

An example of the failure can be seen in the output of the
“Measure test coverage” step of this CI build:
  https://circleci.com/gh/FundingCircle/fc4c/97

Here’s a snippet of it:

```
FAIL in (snap-coords) (fc4c/test_utils.clj:17)
({:spec
 (fspec
  :args
  (cat
   :coords
   (coll-of nat-int? :count 2)
   :to-closest
   nat-int?
   :min-margin
   nat-int?)
  :ret
  :fc4c.spec/coord-string
  :fn
  (fn
   [{:keys [ret args]}]
   (let
    [parsed-ret
     (parse-coords ret)
     {:keys [:to-closest :min-margin]}
     args]
    (every? (fn* [p1__7468#] (>= p1__7468# min-margin)) parsed-ret)))),
 :sym fc4c.integrations.structurizr.express.edit/snap-coords,
 :failure #error {
 :cause Divide by zero
 :via
 [{:type java.lang.ArithmeticException
   :message Divide by zero
   :at [clojure.lang.Numbers divide Numbers.java 163]}]
 :trace
 [[clojure.lang.Numbers divide Numbers.java 163]
  [clojure.core$_SLASH_ invokeStatic core.clj 1019]
  [clojure.core$_SLASH_ invoke core.clj 1012]
  [fc4c.integrations.structurizr.express.edit$round_to_closest invokeStatic fc4c/integrations/structurizr/express/edit.clj 169]
  [fc4c.integrations.structurizr.express.edit$round_to_closest invoke fc4c/integrations/structurizr/express/edit.clj 0]
  [clojure.core$partial$fn__5561 invoke core.clj 2616]
```