Fix the files that Cloverage instruments
I stumbled across https://github.com/cloverage/cloverage/issues/213
wherein I learned that the CLI options listed in the Cloverage README
did not apply to the latest released version but rather to the next
upcoming release. So this upgrades to the snapshot of the upcoming
release, which _does_ support these options — no more complex regex
needed.