Add Cloverage adapter for our DIY test runner runner
I originally did this work because I was hoping it would speed up the
coverage step of our CI jobs, but surprisingly the execution of the
coverage run is not any faster. I had though that Cloverage uses the
clojure.test runner by default, which is single-threaded, and that by
switching to our own test runner runner, which uses Eftest, which is
multi-threaded, well, I thought that would cut the wall clock time
significantly. But somehow that doesn’t seem to be the case? I have no
idea why.

Turns out this is helpful in a different way, though: now that we fixed
(in a3c4c64c0) the test that was *always* failing when measuring
coverage, we should be able to merge the CI job steps that ran the tests
and measured coverage. Doing both was redundant and slow, but necessary
because the coverage step couldn’t be used to fail the build if the
tests failed. Now that we can, we can use the coverage job to both run
the tests *and* measure coverage, which should make our CI jobs ~30
seconds (~15%) faster. So that’s something.

The only downside to merging these steps would have been that without
this custom adapter, Cloverage would have continued to use one of it’s
built-in runner runners instead of ours, and so we would have been able
to use only *one* reporting function, sending coverage results to
*either* stdout or the XML file that CodeCov understands, when we really
want *both*. So using this adapter enables us to use *our* runner runner,
and thereby get the coverage results reported in the various ways we
prefer.