Move the test runner to its own source tree
It was originally in src, but we moved it to test because it was being
covered by Cloverage and therefore included in the measurements of test
coverage, and we didn’t want that.

But then we added a Cloverage adapter into the same folder as the
runner, which works just fine when we’re measuring coverage, but fails
when we’re *just* running the tests, because in that case Cloverage
isn’t in the classpath, but the eftest function `find-tests` will load
every single clj file it finds in the specified dir, recursively. And
since the Cloverage adapter ns requires Cloverage, it fails.

I guess we *could* just include Cloverage as a dependency in the base
:test profile, but that would be lazy; there’s no reason it should be
there; it is not and shouldn’t be a dependency for just running the
tests.

Creating this new source tree for the test runner means it won’t be
instrumented by Cloverage, and the runner and its Cloverage adapter
won’t be loaded by eftest’s find-tests function.

Bada-bing! Problem solved.