Readability improvement to ..structurizr../spec.clj
This file has been hard to read for a good while. I had experimented
with having multiple `ns` forms in this file so as to be able to use
double-colon aliases in the qualified keywords, but I removed those
because they were non-idiomatic and would have seemed like a code smell
to many Clojurists seeing the file for the first time. So for a good
while I just left this file as-is, with the namespaces written out in
full. It was nice and explicit, but not so readable.

This kept itching at me though, so finally today, for some reason, I
tried a few different approaches and eventually found I can just create
whatever namespaces I want — as _Clojure_ namespaces — and then create
aliases to those namespaces. I’m happy with this approach.