Add generative tests to ...structurizr.express.export
This doesn’t actually increase test coverage at the project level,
because we have an example test that calls the big
compose-it-all-together function at the bottom of the namespace. But
this is still helpful because it serves as documentation and regression
tests for these individual functions, so bugs can be found in the future
at a useful granularity.

Also a few code formatting tweaks.