A few minor but important spec tweaks
This isn’t just for fun; I had some functions returning empty versions
of these collections… so I decided this is valid. The question is:
what’s the minimum valid Structurizr Express diagram? I’m thinking it
has to have at least one element, but I suppose relationships and styles
are not absolutely necessary — such a diagram would simply show a single
element in the specified location. But that could hypothetically be
useful. We’ll see!