Add another fspec test and fix bugs it revealed
For some reason I was under the impression that
`clojure.spec.test.alpha/check` would only work on fspec specs that
included a :fn property-checking function. I had been thinking that that
was a shame, and surprising, because I would think there’d be a ton of
value in just generating inputs, passing them to the specified function,
and validating the outputs against the spec. Turns out, yeah, there is,
and `check` supports that just fine. Good to know!