Tweaks some expressions in a function property test
When I first wrote this property test, system references could contain
either :system or :container or both. I subsequently changed the
definition of system references (see model_spec.clj) so that :system is
always present, and :container is optional. This is far easier to reason
about.

This change had the consequence of making the logic in this `or` form
non-sensical — it would check first that the destination was equal to
the :system key — but for now that’s now an invariant, because currently
this `export` namespace is designed only to handle System Context
diagrams, in which _every_ destination is a :system and not a :container.
Only in a :container diagram would a relationship destination be a
container.

So this all together meant that the other branch was never being
evaluated, which reduced test coverage. So by swapping the expressions
around, we can increase test coverage.

I suppose having any code at all that supports or references a container
being a relationship destination violates YAGNI, but not really, because
I know I *am* going to need it — I’m planning to add support
container diagrams very soon.