Fix #50
Turns out this bug was caused by a refactoring I did back in 5ef13deaa
to change process-file to return a map rather than a tuple. This is
“just” an invocation of process-file that I forgot to update, that was
lacking code coverage.

This does improve coverage just a bit (but not enough) as I added a
failing test to ensure that the fix actually works.

Fixes #50