Improve specs for Structurizr Express elements
I first discovered a minor issue with the spec for `element` via
instrumentation — I think it was something to do with the rules around
which elements require `position` and `containers` and when. So I
started trying to make some minor tweaks, e.g. I tried specifying that
an element must contain either `position` or `containers` — but that
didn’t work because then I’d end up with “person” elements with
“containers”. So in the end I realized that I needed to specify each
kind of element individually, and then I could have a big s/or spec
that’d cover all of them, for those functions for which that’s useful.

This works, but given these maps do have a `type` field I wonder whether
it would be better to use multi-spec for this.