Minor spec fix
Bug discovered via spec instrumentation.

Now that I’m thinking about this though, I’m thinking it’s bad that
this bug slipped through, and also bad that the values passed to
process-file and process-spec aren’t checked at runtime; it would be
useful to warn users if the data doesn’t seem correct, and maybe to
abort the operation (process or process-file) as well.

This may be reminding me of some of the limits of generative testing —
the tests are only as good as your specs, and they can’t find bugs in
your specs.

So it may be a good idea to add a few example tests — especially for
process and process-file.