Fix bug in and add spec for ..edit/reorder
Well maybe not _exactly_ a bug; maybe a mistake... anyway, as mentioned
in the comment fragment removed by this change, reorder would sometimes
introduce new keys with nil values into the maps being (re)ordered. This
was fine before we were using spec, but now that we _are_ using spec,
some of those keys don’t allow nil values, and the results of reorder
were therefore invalid (as per the specs).

Adding the spec helped me realize that I don’t want this function to
accept an empty sequence of keys — that wouldn’t make sense. So I wrote
the args spec accordingly, and for good  measure I added a precondition
to the function so that if an empty seq is passed to it, it’ll throw an exception.