Remove keyword namespaces in ..edit/snap-elem-to-grid
These keywords were qualified in e64ab931e (the namespace was
subsequently changed) but doing so was erroneous — these keywords should
have remained “simple”.

The reason is: the data structure this function operates on is a
Structurizr Express diagram, which is a map with unqualified keyword
keys. By adding namespaces to the keywords used here as “getters” we
ensured that these “getters” always returned `nil` and therefore this
function essentially ceased to do its job with that change — yeah, back
in April 🤦‍♂️.

While I was fixing that, I noticed that the predicate in the :fn slot of
the s/fdef made no sense — so I had to fix it. It was a little tricky to
disentangle these changes so they ended up in this commit together.