Fix some test data
Now that the snap-to-grid feature was fixed (in the prior commit) this
test data needed to change, because it’s intended to pass through
process-file unchanged, but it hadn’t been accounting for the offsets
that get applied to elements of type `Person`. Now it does, and so now
it once again passes through `process-file` unchanged.

I added the style just so if/when this test diagram is rendered, it
looks right. Without this, the Person element wouldn’t actually align
with the System element. Quirk of Structurizr Express.