Add missing generative test and fix spec thereof
I thought this would be as simple as adding the call to `check` but of
course it turns out that the spec was mostly bad and had to be
rewritten. And I got real lost doing that, as I’m *still* not used to
working with values that have been conformed by spec, and I got bit by
the fact that both the values in :args and :ret passed to the :fn
predicate were conformed.

And *then* I had trouble debugging *that*
because I got bit by the [trace missing] problem:
  https://theholyjava.wordpress.com/2014/05/19/clojurejava-prevent-exceptions-with-trace-missing/