Tool: update some test deps
I had been hoping to be able to remove the :multithreading? option from
the eftest config, in the hopes that the default would be simpler,
faster, and have lower maintenance implications (cognitive load).

I had thought this might be possible because the latest version of
Eftest includes this fix by @nachomdo:
  https://github.com/weavejester/eftest/pull/58

Unfortunately when I tested the other possible values for
:multithreading? including `true` which is the default, the tests took
significantly longer to run (on my laptop). I’m not sure why, but
regardless, that’s no good.

Given these results, this change doesn’t seem to yield any compelling
concrete benefits right now; rather by updating the comment to remove
the reference to the bug that Nacho fixed, hopefully this will obviate
the need for someone else to think about this in the future. i.e.
potentially slightly lower future maintenance costs.

In addition I think it’s just a generally and broadly beneficial
practice to keep dependencies up to date when practical. So there’s that
at least.