Add link to the "fc4-tool"
We have a link to the
https://fundingcircle.github.io/fc4-framework/methodology/toolset.html
page directly.

New users learn about the "fc4-tool" but cannot see what it is unless
they navigate away from the page.

This change adds a link to the directory containing the tool in GitHub
to make resolving "fc4-tool" less of a quest.