[tool: automated rendering] Still trying to fix CI
I’ve been having a lot of trouble getting Puppeteer and Chromium to work
in the Alpine CI image… finally I found this:

https://github.com/GoogleChrome/puppeteer/blob/master/docs/troubleshooting.md#running-on-alpine

which informed me that right now we have to use Puppeteer 1.4.0 (as
opposed to the current version, 1.9.0). And it also recommended using
the apk-installed Chromium rather than having Puppeteer download its own
version, which makes sense to me since we need to have apk install
chromium either way, so as to get all the dependencies installed that we
need to run any chromium.