[tool: automated rendering] Restore Structurizr logo to rendered images
By adjusting the duration of a pause in rendering process.

The first change, the increase, is the one that restored the logo; I
realized that a prior change that shortened this pause had caused the
logo to be missing.

The second change, the decrease, I tried in the hopes of offsetting the
first. It seems to work on my system, but might need to be bumped back
up. We’ll see.