[tool: automated rendering] Fix intermittent issue with missing logo
I recently observed that the Structurizr Express logo was always missing
from the rendered images, and that lengthening the duration of a certain
pause in the rendering process fixed this. It turns out I didn’t
lengthen it quite enough, because I’m now seeing the logo be
occasionally, intermittently, missing.

So hopefully bumping this from 300ms to 350ms will make this issue less
likely to occur.

There may be smarter ways to ensure the image is on the
diagram, though; I think maybe listening for the `load` event on the
image object in the DOM might be faster _and_ more accurate a way to
both ensure that the logo appears _and_ that we’re pausing only for the
absolute minimum necessary duration.