[tool: automated rendering] Make renderer work with remote Structurizr Express
I realized that we can’t embed Structurizr Express (SE) into this
project unless and until we get permission from Simon Brown and work out
licensing issues. So I’ve broken that work (embedding SE) out into a
separate branch, so this renderer for now at least needs to work with
the canonical instance of Structurizr Express.

While working on this, I found I needed some debugging help, so I added
a debugMode to help out with that. And while trying to get the timing of
the pauses right, I discovered that Puppeteer has a page.waitForSelector
function, which was perfect for replacing one of the sleeps with
something more precise and probably more reliable.