[tool] Add support for creating an AOT-compiled Uberjar
Proximate catalyst is to address #85 wherein the project cannot
(currently) be compiled under Java 11 due to an incompatibility between
Java 11 and one of the project’s indirect dependencies (flatland/ordered
via clj-yaml).

By AOT compiling the project on Java <11 and then packaging up the
compiled code along with its dependencies, the resulting JAR will run
just fine on Java 11. It’s not quite as flexible as running the project
from source using `clojure` -- the “entrypoint” (main class) is
hard-coded -- but for now that’s the only command that’s really
relevant to end-users.

This also makes some progress on
[Enhancement Proposal EP04: Improve the Initial UX of fc4-tool][ep04]
by removing the requirements that end-users clone the project and
install Clojure; now they just need Java and they can download the
packaged JAR and run it.

TODO: eliminate the reflection warnings that are printed out when
running the jar under Java 11.

TODO: move the process of building this JAR and uploading it to GitHub
to CI.

[ep04]: https://github.com/FundingCircle/fc4-framework/pull/76