[tool: automated rendering] Attempt to fix render test
The problem is that Structurizr Express (SE) puts its watermark in a
random location at a random angle, so I can’t expect its output to be
exactly identical to any kind of pre-saved “expected” value.

So this attempt uses an approach of blowing out the contrast of both
images and hoping that that makes them exactly equal.

There’s also some remnants of another approach in here, using a FilteredImageSource with a filter to remove the watermark’s pixels by
color-matching… but I couldn’t get that to work. Both approaches will
be removed soon anyway so they’re really just included in the change
history for historical reference.