[tool][renderer] use alternate approach to reading stdin
I can’t use for await..of because that’s not currently supported by
pkg[1] and it’s not worth trying to patch pkg or use a fork or
whatever... trust me, I tried.

Using readFileSync is an unusual way of reading a stream in Node, I
think, but I had all sorts of trouble trying to do it the right way,
with event handlers. I don’t know, maybe something to do with TTY, but
I don’t care. I’m really not fond of Node right now.

[1] https://github.com/zeit/pkg/pull/538