[tool] Start WIP: package renderer with pkg
I’ve given up on bundling Chromium in the executable, it’s just too
tricky/iffy; see https://github.com/zeit/pkg/issues/204

Now that I think of it, this is an end-user app and most users will
already have Chrom[e|ium] installed, so maybe it’s better anyway
not to bundle Chrom[e|ium] and make the download longer and use more of
the user’s disk space.

So this attempts to find the location of the user’s Chrom[e|ium] on the
fly at runtime and use it. This needs shoring up.

WIP++ attempt to package entire tool, uberjar+renderer+script