[tool] Fix CircleCI publish artifact job
In the prior commit (e657c47610) I changed the workflow and job so that
the publish job will be run for all builds on the master branch, rather
than on tag builds. I clearly forgot to remove this check that requires
that $CIRCLE_TAG be set and non-blank, because with this check in place
and the workflow only running the job for master branch builds, this job
will never work.