[tool] Update requirements in docs
I discovered today that the version of Puppeteer we’re currently using
(1.4.0) no longer works with the latest Chromium/Chrome (72) — it
requires Chromium/Chrome 68 (it _might_ work with 69–71 but they’re
explicitly not supported).

This is annoying, and we’re going to have to do something about it soon,
but in the meantime I think it’s necessary to explain this requirement
and provide some assistance in getting the required version installed.

Related: https://github.com/FundingCircle/fc4-framework/issues/92