[tool] Support recent/current versions of Chrome/Chromium
This switches our test+build containers from Alpine to Debian Stretch,
which has a more recent version of Chromium. This allows us to upgrade
to a less ancient version of Puppeteer; this less ancient Puppeteer is
compatible with modern Chrome/Chromium. Therefore users no longer need
to have the ancient+obsolete Chrome/Chromium 68 installed on their
systems, which was unlikely, annoying, and insecure.

This also upgrades from Node 10 to Node 11, because why not 😬

Fixes #92