[tool ci] Separately download+cache deps for testing+packaging
I’ve realized that for the Clojure side of things at least (which is
most of the tool) it’s important to download the testing deps and the
packaging deps separately, completely discretely, because otherwise we
get a notably different end-result.

So I’ve rearranged the scripts in tool/bin and the CI jobs accordingly.

It no longer makes sense to have a single download-deps job that would
cache the deps for _all_ the subsequent jobs; each job has a different
set of dependencies that it needs. I’ve therefore removed that job and
now the test, package, and lint jobs manage their own caches, as is more
conventional with CircleCI jobs anyway.

I think this’ll make the jobs faster, and more robust, and *might* make
the Clojure distribution package a little smaller.