[tool] Switch from puppeteer to puppeteer-core
The only difference between these libs is that puppeteer downloads a
guaranteed-compatible version of Chromium into a local subdirectory
(under `tool/renderer/node_modules/puppeteer/`) while puppeteer-core
does not — when using puppeteer-core one must pass the path to a
Chromium or Chrome executable to the library when calling `launch` to
launch a browser.

Well, in our case we don’t want to download an “embedded” Chromium to be
included in `node_modules`, for two reasons:

1. It’d make the distribution package massive; possibly an extra ~100MB
   or more
   * And this’d be pretty wasteful, as most of our users will probably
     already have a compatible Chrome or Chromium installed
2. pkg can’t or won’t include that automatically-downloaded version of
   Chromium
   * See https://github.com/zeit/pkg/issues/204

Because of the packaging problem, we’re already passing the path to a
system-level Chromium or Chrome executable to the library when calling
`launch` to launch a browser, and we’ve been setting an
environment variable to direct puppeteer _not_ to download Chromium.
Therefore it’s silly to use puppeteer rather than puppeteer-core, as
it’s actually more work and more complex.

Even though this is complicated to explain, I think it makes the tool
a tiny bit simpler.