[tool] Ensure renderer is never published to npm
Partly because it wouldn’t even work, as there’s already a package with
this name published in npm. Mainly because this is not meant to be a
standalone library or tool; it’s only ever meant to be run from source
from within the context of this repository.

As pointed out by @sgerrand in #95 — thanks Sasha!