[tool] Fix + improve automated rendering
Simon Brown, the owner and maintainer of Structurizr Express, has made
some changes to its code that broke our renderer.

Thankfully, he also published a new example script[1] that demonstrates
the optimal way to render and export diagrams via Structurizr Express;
this enabled me to both fix our renderer but also improve it.

So this adapts the renderer to the changes but also makes our code
simpler and faster.2

[1] https://github.com/structurizr/puppeteer