[tool] Resolve a dependency conflict
In the prior commit I decoupled the downloading of dependencies for the
test job and the packaging job. This worked for the tests with no issues
but it revealed that when the uberjar profile was used on its own, it
resulted in a dependency conflict that prevented the uberjar process
from succeeding.

Specifically, after Cambada compiled all the Clojure source files in the
project, and started building the JAR, it’d output this:

```
Creating target/tool-1.0.0-SNAPSHOT.jar
Updating pom.xml
Skipping paths: src/main src/cli
Exception in thread "main" org.eclipse.aether.resolution.ArtifactResolutionException: Could not find artifact joda-time:joda-time:jar:2.7
```

I wasn’t sure at first how to identify the source of the conflict, but
then I realized I could run `clojure -A:test -Stree` and compare its
output with that of `clojure -A:uberjar -Stree` and voila:

the 2.8.1 dependency comes from cambada, which depends on s3-wagon-private:

```
    s3-wagon-private/s3-wagon-private 1.3.1
      org.springframework.build/aws-maven 4.8.0.RELEASE
      com.fasterxml.jackson.core/jackson-databind 2.5.5
        com.fasterxml.jackson.core/jackson-annotations 2.5.0
      com.fasterxml.jackson.core/jackson-core 2.5.5
      com.amazonaws/aws-java-sdk-s3 1.11.184
        com.amazonaws/jmespath-java 1.11.184
        com.amazonaws/aws-java-sdk-core 1.11.184
          joda-time/joda-time 2.8.1
```

and the 2.7 dependency:

```
com.gfredericks/test.chuck 0.2.9
  clj-time/clj-time 0.10.0
    joda-time/joda-time 2.7
```