[tool renderer] Change url argument to a constant
Because it doesn’t actually ever vary.

The only reason it was an argument is because at one point I was
experimenting with embedding Structurizr Express into the renderer and
loading it from the filesystem — but that turned out to be a non-viable
option.

Prompted by a suggestion from @fussybeaver on #106 — thanks Niel!