[tool] Remove requirement that renderer be in user’s $PATH
By figuring out the renderer’s absolute path by assuming that it’ll be
in the same dir as the uberjar!

This way, `fc4` will be usable like any standard UNIX program; it’ll be
runnable no matter where it is; it could be run in-place.

To illustrate:

```shell
# Let’s say we download the archive and extract it
curl -L -s $URL | tar xzv

# Before this change, this wouldn’t work:
./fc4/fc4 render path/to/my/diagram.yaml

# Because `fc4` would be invoking `fc4-render` “bare”, hoping that it’d
# be in the user’s path. So it’d have to be.

# But now it will work.

# I think.
```

Closes #105