[tool] Fix bug: renderer-command was returning a File object
When the tool was run from within its uberjar. It’s supposed to return
a string.

The commit with the bug passed CI, which inspired me to tweak the CI
test render to use the regular `fc4` wrapper script and thereby test the
entire workflow of `fc4 render` just like a regular user would.